+++ CHANGELOG.md
@@ -3,6 +3,7 @@ UNRELEASED
 
 * [ [#1271](https://github.com/digitalfabrik/integreat-cms/issues/1271) ] Fix feedback API endpoint
 * [ [#1099](https://github.com/digitalfabrik/integreat-cms/issues/1099) ] Add push content API
+* [ [#1277](https://github.com/digitalfabrik/integreat-cms/issues/1277) ] Fix change detection for XLIFF import
 
 
 2022.3.2